amd-power-cap: Add flag to indicate APML init complete
APML slaves are behind i3c mux in CRB. i3c mux is powered on
when host rails are on. Due to this design, we need to bind APML drivers
twice. First, to configure mux and second to configure APML slave.

This change adds flag (file) to indicate that binding of APML i3c driver
was successful and APML slaves are configured correctly.

Phosphor-pid-control shall wait for this flag (ConditionPathExists) and
then query APML based sensors (CPU temp)

Signed-off-by: Rajaganesh Rathinasabapathi <Rajaganesh.Rathinasabapathi@amd.com>